Move & modify signup page component
